move group tools to own file and refactor highest level collapse to use items
